+++ passkit/lib/src/pkpass/pkpass.dart
@@ -1,24 +1,17 @@
-import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:archive/archive.dart';
-import 'package:crypto/crypto.dart';
 import 'package:meta/meta.dart';
+import 'package:passkit/src/archive_extensions.dart';
 import 'package:passkit/src/archive_file_extension.dart';
+import 'package:passkit/src/pk_image.dart';
 import 'package:passkit/src/pkpass/exceptions.dart';
 import 'package:passkit/src/pkpass/pass_data.dart';
 import 'package:passkit/src/pkpass/pass_type.dart';
 import 'package:passkit/src/pkpass/personalization.dart';
-import 'package:passkit/src/pkpass/pk_pass_image.dart';
 import 'package:passkit/src/signature_verification.dart';
-import 'package:passkit/src/strings_parser/naive_strings_file_parser.dart';
+import 'package:passkit/src/utils.dart';
 import 'package:passkit/src/write_signature.dart';
-
-/// Dart uses a special fast decoder when using a fused [Utf8Decoder] and [JsonDecoder].
-/// This speeds up decoding.
-/// See
-/// - https://github.com/dart-lang/sdk/blob/5b2ea0c7a227d91c691d2ff8cbbeb5f7f86afdb9/sdk/lib/_internal/vm/lib/convert_patch.dart#L40
-final _utf8JsonDecoder = const Utf8Decoder().fuse(const JsonDecoder());
 
 /// https://developer.apple.com/library/archive/documentation/UserExperience/Reference/PassKit_Bundle/Chapters/Introduction.html
 /// https://developer.apple.com/library/archive/documentation/UserExperience/Conceptual/PassKit_PG/index.html#//apple_ref/doc/uid/TP40012195
@@ -107,13 +100,13 @@ class PkPass {
       // TODO(ueman): Images can be localized, too
       //              Maybe it's better to have an on-demand API, something like
       //              PkPass().getLogo(resolution: 3, languageCode: 'en_EN').
-      logo: archive.loadImage('logo'),
-      icon: archive.loadImage('icon'),
-      footer: archive.loadImage('footer'),
-      thumbnail: archive.loadImage('thumbnail'),
-      strip: archive.loadImage('strip'),
-      background: archive.loadImage('background'),
-      personalizationLogo: archive.loadImage('personalizationLogo'),
+      logo: archive.loadPkPassImage('logo'),
+      icon: archive.loadPkPassImage('icon'),
+      footer: archive.loadPkPassImage('footer'),
+      thumbnail: archive.loadPkPassImage('thumbnail'),
+      strip: archive.loadPkPassImage('strip'),
+      background: archive.loadPkPassImage('background'),
+      personalizationLogo: archive.loadPkPassImage('personalizationLogo'),
       // source
       sourceData: bytes,
     );
@@ -159,7 +152,7 @@ class PkPass {
 
   /// Mapping of files to their respective checksums. Typically not relevant for
   /// users of this package.
-  final Map<String, dynamic> manifest;
+  final Map<String, dynamic>? manifest;
 
   /// The [PassType] of this PkPass.
   PassType get type {
@@ -178,7 +171,7 @@ class PkPass {
     if (pass.storeCard != null) {
       return PassType.storeCard;
     }
-    return PassType.unknown;
+    throw Exception('unknown pass type');
   }
 
   /// The image displayed as the background of the front of the pass.
@@ -227,7 +220,8 @@ class PkPass {
   final Map<String, Map<String, dynamic>>? languageData;
 
   /// The bytes of this PkPass
-  final Uint8List sourceData;
+  /// Returns `null` when this instance wasn't read from a file.
+  final Uint8List? sourceData;
 
   /// Indicates whether a webservices is available.
   bool get isWebServiceAvailable => pass.webServiceURL != null;
@@ -259,23 +253,24 @@ class PkPass {
     String? privateKeyPem,
   }) {
     final archive = Archive();
-    final encoder = JsonEncoder.withIndent('  ');
 
-    final passFile = ArchiveFile.string(
+    final passContent = utf8JsonEncode(pass.toJson());
+    final passFile = ArchiveFile(
       'pass.json',
-      encoder.convert(pass.toJson()),
+      passContent.length,
+      passContent,
     );
     archive.addFile(passFile);
-    /*
+
     if (personalization != null) {
-      encoder.addFile(
-        ArchiveFile.string(
-          'personalization.json',
-          jsonEncode(personalization!.toJson()),
-        ),
+      final personalizationContent = utf8JsonEncode(personalization!.toJson());
+      final personalizationFile = ArchiveFile(
+        'personalization.json',
+        personalizationContent.length,
+        personalizationContent,
       );
+      archive.addFile(personalizationFile);
     }
-    */
 
     logo?.writeToArchive(archive, 'logo');
     background?.writeToArchive(archive, 'background');
@@ -283,24 +278,18 @@ class PkPass {
     footer?.writeToArchive(archive, 'footer');
     strip?.writeToArchive(archive, 'strip');
     thumbnail?.writeToArchive(archive, 'thumbnail');
+    personalizationLogo?.writeToArchive(archive, 'personalizationLogo');
 
-    final manifest = <String, String>{};
-    for (final file in archive.files) {
-      manifest[file.name] = sha1.convert(file.binaryContent).toString();
-    }
-
-    final manifestContent = encoder.convert(manifest);
-    final manifestFile = ArchiveFile.string(
-      'manifest.json',
-      manifestContent,
-    );
-    archive.addFile(manifestFile);
+    final manifestFile = archive.createManifest();
 
     if (certificatePem != null && privateKeyPem != null) {
       final signature = writeSignature(
         certificatePem,
         privateKeyPem,
-        Uint8List.fromList(manifestFile.content as List<int>),
+        manifestFile,
+        pass.passTypeIdentifier,
+        pass.teamIdentifier,
+        true,
       );
 
       final signatureFile = ArchiveFile(
@@ -312,19 +301,13 @@ class PkPass {
     }
 
     final pkpass = ZipEncoder().encode(archive);
-    if (pkpass == null) {
-      return null;
-    }
-    return Uint8List.fromList(pkpass);
+    return pkpass == null ? null : Uint8List.fromList(pkpass);
   }
 }
 
 // This is intentionally not exposed to keep this an implementation detail.
 // Tests should be written against the PkPass class directly.
 extension on Archive {
-  Uint8List? findBytesForFile(String fileName) =>
-      findFile(fileName)?.binaryContent;
-
   /// Returns a map of locale to a map of resolution to image bytes.
   /// Returns null, if no image is localized
   Map<String, Map<int, Uint8List>>? loadLocalizedImage(String imageName) {
@@ -375,34 +358,16 @@ extension on Archive {
     if (bytes == null) {
       return null;
     }
-    return _utf8JsonDecoder.convert(bytes) as Map<String, dynamic>?;
+    return utf8JsonDecode(bytes);
   }
 
-  PkImage? loadImage(String name) {
+  PkImage? loadPkPassImage(String name) {
     return PkImage.fromImages(
       image1: findBytesForFile('$name.png'),
       image2: findBytesForFile('$name@2x.png'),
       image3: findBytesForFile('$name@3x.png'),
       localizedImages: loadLocalizedImage(name),
     );
-  }
-
-  Map<String, Map<String, String>> getTranslations() {
-    final languageData = <String, Map<String, String>>{};
-
-    // The Archive object doesn't have APIs to work with folders.
-    // Instead the file name contains a `/` indicating the file is within a folder.
-    // Example: `file.name == en.lproj/pass.strings`
-    final translationFiles = files
-        .where((element) => element.isFile)
-        .where((file) => file.name.endsWith('.lproj/pass.strings'));
-
-    for (final languageFile in translationFiles) {
-      final language = languageFile.name.split('.').first;
-
-      languageData[language] = parseStringsFile(languageFile.binaryContent);
-    }
-    return languageData;
   }
 
   PassData readPass() {
@@ -414,49 +379,12 @@ extension on Archive {
     }
   }
 
-  Map<String, dynamic> readManifest() {
-    final manifestJson = findFileAndReadAsJson('manifest.json');
-    if (manifestJson != null) {
-      return manifestJson;
-    } else {
-      throw MissingManifestJsonException();
-    }
-  }
-
   Personalization? readPersonalization() {
     final personalizationFile = findFileAndReadAsJson('personalization.json');
     if (personalizationFile != null) {
       return Personalization.fromJson(personalizationFile);
     }
     return null;
-  }
-
-  /// To create the manifest file, recursively list the files in the package
-  /// (except the manifest file and signature), calculate the SHA-1 hash of the
-  /// contents of those files, and store the data in a dictionary. The keys are
-  /// relative paths to the file from the pass package. The values are the SHA-1
-  /// hash (hex encoded) of the data at that path. Write this dictionary to the
-  /// file manifest.json at the top level of the pass package.
-  ///
-  /// https://developer.apple.com/library/archive/documentation/UserExperience/Conceptual/PassKit_PG/Creating.html#//apple_ref/doc/uid/TP40012195-CH4-SW1
-  void checkSha1Checksums(Map<String, dynamic> manifest) {
-    final filesWithoutSignatureAndManifest = files.where((file) {
-      return file.name != 'signature' && file.name != 'manifest.json';
-    }).toList();
-
-    final fileInArchiveCount = filesWithoutSignatureAndManifest.length;
-    final manifestCount = manifest.length;
-    if (fileInArchiveCount != manifestCount) {
-      throw MissingChecksumException();
-    }
-
-    for (final file in filesWithoutSignatureAndManifest) {
-      final checksumInManifest = manifest[file.name] as String?;
-      final digest = sha1.convert(file.binaryContent);
-      if (checksumInManifest != digest.toString()) {
-        throw ChecksumMismatchException(file.name);
-      }
-    }
   }
 }