+++ passkit/test/pkpass/create_pass_test.dart
@@ -29,8 +29,8 @@ void main() {
     );
 
     // pass.json file is not identical due to whitespace differences
-    final copyManifest = copy.manifest..remove('pass.json');
-    final ogManifest = og.manifest..remove('pass.json');
+    final copyManifest = copy.manifest?..remove('pass.json');
+    final ogManifest = og.manifest?..remove('pass.json');
 
     expect(copyManifest, ogManifest);
   });