+++ passkit/CHANGELOG.md
@@ -1,10 +1,10 @@
-## Unreleased
+## 0.0.8
 
 - Make it possible to override the bundled Apple WWDR G4 certificate
   - Use the `overrideWwdrCert` argument of `PkPass.fromBytes()`
   - Use the `overrideWwdrCert` argument of `PkOrder.fromBytes()`
 - Added the ability to create localized passes
-- Remove dependencies on the `encrypt` and `asn1lib` packages
+- Remove dependencies on the `encrypt` and `asn1lib` packages.
 
 ## 0.0.7